Discovered that i need to port libunwind
Damn it's a mess to port. In order to build the libcxxabi, I need to
have libunwind ported. But that library is a damn mess.

+++ librt/include/os/driver/input.h
@@ -52,9 +52,9 @@ typedef struct _MInput {
 	unsigned			Scancode;
 	VKey				Key;
 	unsigned			Flags;		/* Flags (Bit-field, see under structure) */
-	ssize_t				xRelative;
-	ssize_t				yRelative;
-	ssize_t				zRelative;
+	long 				xRelative;
+	long				yRelative;
+	long				zRelative;
 } MInput_t;
 
 /* Flags - Event Types */
